(cherry-pick): Fix signed-unsigned-compare warning (reported by François Vogel on Windows, but gcc can trigger it as well)

+++ generic/tclCompExpr.c
@@ -666,7 +666,7 @@ ParseExpr(
 	    OpNode *newPtr = NULL;
 
 	    do {
-	      if (size <= UINT_MAX/sizeof(OpNode)) {
+	      if (size <= (int)(UINT_MAX/sizeof(OpNode))) {
 		newPtr = attemptckrealloc(nodes, size * sizeof(OpNode));
 	      }
 	    } while ((newPtr == NULL)
